added client call GetAllPools

+++ Source/Sprd/Cardano-Wallet/Model/StakepoolsCost.cs
@@ -50,7 +50,7 @@ namespace IO.Swagger.Model
         /// </summary>
         /// <param name="quantity">quantity (required).</param>
         /// <param name="unit">unit (required).</param>
-        public StakepoolsCost(int? quantity = default(int?), UnitEnum unit = default(UnitEnum))
+        public StakepoolsCost(Int64? quantity = default(Int64?), UnitEnum unit = default(UnitEnum))
         {
             // to ensure "quantity" is required (not null)
             if (quantity == null)
@@ -76,7 +76,7 @@ namespace IO.Swagger.Model
         /// Gets or Sets Quantity
         /// </summary>
         [DataMember(Name="quantity", EmitDefaultValue=false)]
-        public int? Quantity { get; set; }
+        public Int64? Quantity { get; set; }
 
 
         /// <summary>
